% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_ci.R
\name{corr_ci}
\alias{corr_ci}
\title{Creates a correlation table with associated confidence intervals}
\usage{
corr_ci(data, digits = 2, export = FALSE, sep = ";")
}
\arguments{
\item{data}{Data frame.}

\item{digits}{Number of digits to print.}

\item{export}{Logical. If TRUE, the correlation table is exported as a csv file to the working directory. The default is FALSE.}

\item{sep}{The field separator string to be used if export = TRUE.}
}
\value{
Correlation matrix with bootstrapped 95\% confidence intervals.
}
\description{
Creates a correlation matrix with associated bootstrapped 95\% confidence invervals for easy export to Excel.
Based on psych::cor.ci().
}
