% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_se.R
\name{corr_se}
\alias{corr_se}
\title{Creates a correlation table with associated standard errors}
\usage{
corr_se(data, digits = 2, export = FALSE, sep = ";")
}
\arguments{
\item{data}{Data frame.}

\item{digits}{Number of digits to print.}

\item{export}{Logical. If TRUE, the correlation table is exported as a csv file to the working directory. The default is FALSE.}

\item{sep}{The field separator string to be used if export = TRUE.}
}
\value{
Correlation matrix with SEs.
}
\description{
Creates a correlation matrix with associated SEs for easy export.
}
