% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_idiff.R
\name{plot_idiff}
\alias{plot_idiff}
\title{Plot proportion correct (difficulty) scores of dichotomous variables}
\usage{
plot_idiff(data, guess = 0.25, sort = "orig", group = NULL)
}
\arguments{
\item{data}{Either a data frame containing (only) dichotomous variables or the output of psych::describe()}
}
\value{
A plot of the means with associated errors bars (2*SE)
}
\description{
This function loads or creates an object from psych::describe() and plots the
mean variable with the associated error bars. The function can easily be
extended with ggplot2 commands.
}
