% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_idiff_paired.R
\name{plot_idiff_paired}
\alias{plot_idiff_paired}
\title{Plot proportion correct (difficulty) scores of dichotomous variables in two groups}
\usage{
plot_idiff_paired(
  data1,
  data2,
  guess = 0.25,
  sort = "orig",
  group = NULL,
  obs = FALSE
)
}
\arguments{
\item{data1}{Either a data frame containing (only) dichotomous variables or the output of psych::describe() of group (or itemset) 1.}

\item{data2}{Either a data frame containing (only) dichotomous variables or the output of psych::describe() of group (or itemset) 2.}

\item{guess}{Optional. Indicate the (theoretical) guessing probability.}

\item{sort}{Optional. Keep variables in the same order as in the dataset (default), order them ascending order ("asc") or or
or order them in descending order ("desc").}

\item{group}{Optional. Character vector as long as the number of variables indicating their group. Item form different
groups will have different colors in the plot.}

\item{obs}{Optional. Decides whether the number of observations is printed below the item label on the x-axis (Default: FALSE).}
}
\value{
A plot of the means with associated errors bars (1*SE)
}
\description{
This function loads or creates two objects from psych::describe() and plots the
mean variables with the associated error bars. The function can easily be
extended with ggplot2 commands.
}
